+++ R/sys.R
@@ -2,5 +2,5 @@
 .Lmax <- 1e+04
 
 .parameters <- c("tmax", "k", "Linf", "t0", "Wb", "Lm", "fb", "Rt",
-                "S", "Lv", "Vp", "Llo", "Lup", "Nc", "mu", "rho", "eta", "Rk",
+                "S", "Mb", "Lv", "Vp", "Llo", "Lup", "Nc", "mu", "rho", "eta", "Rk",
                 "R0", "Wa", "fa")