+++ R/utils.R
@@ -9,3 +9,7 @@ quantiles <- function(x, level) {
   x <- stats::quantile(x, c(0.5, (1-level)/2, (1-level)/2 + level))
   setNames(x, c("estimate", "lower", "upper"))
 }
+
+instantaneous_to_interval <- function(x) 1-exp(-x)
+
+interval_to_instantaneous <- function(x) -log(1-x)